Added 5.4.2

+++ pull.sh
@@ -2,7 +2,7 @@
 
 if [ $# == 0 ]
 then
-	echo "usage \"./pull52.sh <targetdir>\" (./pull52.sh ../52/)"
+	echo "usage \"./pull.sh <targetdir>\" (./pull.sh ../paparazzi_dir/)"
 else
 	if [ -z `dpkg -l | grep paparazzi-dev | cut -c5-17` ]
 	then	
@@ -24,11 +24,32 @@ else
 		fi
 	fi
 
+        read -p"Do you want to install Paparazzi 5.2 stable (1) or 5.4.2 stable (2)?" response
+
+        if [ "$response" == "1" ]; then
+                version="52"
+        elif [ "$response" == "2" ]; then
+		version="54"
+	else
+                exit
+        fi
+
+	if [ "$version" == "52" ]; then
 	
-	git clone https://github.com/paparazzi/paparazzi.git $1
-	origin=`pwd`
-	cd $1
-	git reset --hard "v5.2.0_stable"	
+		git clone https://github.com/paparazzi/paparazzi.git $1
+		origin=`pwd`
+		cd $1
+		git reset --hard "v5.2.0_stable"	
+	fi
+
+        if [ "$version" == "54" ]; then
+        
+                git clone https://github.com/paparazzi/paparazzi.git $1
+                origin=`pwd`
+                cd $1
+                git reset --hard "v5.4.2_stable"
+        fi
+
 
 	read -p"do you want me to patch your Paparazzi (y/n)? " response
 
@@ -44,6 +65,7 @@ else
 	if [ "$response" == "y" ]; then
 	        cd $origin
 	        cd $1
+		. ENV
 		cd sw/ground_segment/lpc21iap
 		make
 		cd $origin
